Converted ship_via PK to UUID
Converted address PK to UUID

Changed vendor_payment_term pk to UUID

Updated referencing columns of vendor_payment_term.id to be called vendor_payment_term_id

Updated vendor_payment_term_schedule pk to be UUID

Changed vendor_group pk to uuid

Changed vendor pk to UUID

changed vendor pk to UUID

Updated account pk to UUID

Updated bank to use UUID for pk

updated account_payable_control to use UUID as the pk

updated purchase_order_header to use UUID as the pk

rebate to use UUID as the pk

purchase_order_header_status_action to use UUID as the pk

bank_reconciliation to use UUID as the pk

general_ledger_control to use UUID as the pk

general_ledger_source_codes to use UUID as the pk

account_payable_invoice to use UUID as the pk

account_payable_invoice_distribution to use UUID as the pk

general_ledger_detail to use UUID as the pk

general_ledger_detail to use UUID as the pk

changed code to use UUID's for entities that can use it

added effective dates

Converted a bunch of code to using UUID's

Converted a bunch of code to using UUID's

Modifyied company entity and related classes

Moved the uuid migration to v41

Changed company primary key to uuid

changed to using the new update_user_table_fn trigger

migrated company and all referencing tables to use UUID for primary key

Modified divsion, region audit related tables to using UUID

changed general_ledger_summary trigger to use the new update function

Organized older table migrations

removed unique constraint on purchase_order_detail

Altered audit_inventory

Fixed AuditDetailControllerSpecification

Updated schedule_arg table to use UUID

Fixed AuditScheduleControllerSpecification

Fixed AuditPermissionControllerSpecification

Fixed RoutineController

Changed general_ledger_control.default_profit_center_sfk to bigint to match the other sfk's in the project

Fixed GeneralLedgerDetail tests

Fixed division and region controllers

Fixed PO tests

Fixed AccountPayableInvoiceControllerSpecification

Fixed Purchas Order Detail tests

Fixed General Ledger Journal Control tests

Fixd audit exception controller specification

Fixed menu tests

Fixed audit tests

Added a new max function that works on UUID's

Formatting cleanup

Fixed docker-compose cynergibasedb context reference

Set jvm max memory to 2 gigs

+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/infrastructure/AccountStatusTypeRepository.kt
@@ -56,7 +56,7 @@ class AccountStatusTypeRepository @Inject constructor(
 
    fun mapRow(rs: ResultSet, columnPrefix: String = EMPTY): AccountStatusType =
       AccountStatusType(
-         id = rs.getLong("${columnPrefix}id"),
+         id = rs.getInt("${columnPrefix}id"),
          value = rs.getString("${columnPrefix}value"),
          description = rs.getString("${columnPrefix}description"),
          localizationCode = rs.getString("${columnPrefix}localization_code")
