+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/infrastructure/NormalAccountBalanceTypeRepository.kt
@@ -56,7 +56,7 @@ private class NormalAccountBalanceCodeTypeRowMapper(
 
    fun mapRow(rs: ResultSet, columnPrefix: String): NormalAccountBalanceType =
       NormalAccountBalanceType(
-         id = rs.getLong("${columnPrefix}id"),
+         id = rs.getInt("${columnPrefix}id"),
          value = rs.getString("${columnPrefix}value"),
          description = rs.getString("${columnPrefix}description"),
          localizationCode = rs.getString("${columnPrefix}localization_code")
