+++ src/main/kotlin/com/cynergisuite/middleware/purchase/order/type/infrastructure/DefaultPurchaseOrderTypeRepository.kt
@@ -54,7 +54,7 @@ class DefaultPurchaseOrderTypeRepository @Inject constructor(
 
    fun mapRow(rs: ResultSet, columnPrefix: String = EMPTY): DefaultPurchaseOrderType =
       DefaultPurchaseOrderType(
-         id = rs.getLong("${columnPrefix}id"),
+         id = rs.getInt("${columnPrefix}id"),
          value = rs.getString("${columnPrefix}value"),
          description = rs.getString("${columnPrefix}description"),
          localizationCode = rs.getString("${columnPrefix}localization_code")
