+++ src/main/kotlin/com/cynergisuite/middleware/shipping/freight/onboard/infrastructure/FreightOnboardTypeRepository.kt
@@ -51,7 +51,7 @@ class FreightOnboardTypeRepository @Inject constructor(
 
    fun mapRow(rs: ResultSet, columnPrefix: String = EMPTY): FreightOnboardType =
       FreightOnboardType(
-         id = rs.getLong("${columnPrefix}id"),
+         id = rs.getInt("${columnPrefix}id"),
          value = rs.getString("${columnPrefix}value"),
          description = rs.getString("${columnPrefix}description"),
          localizationCode = rs.getString("${columnPrefix}localization_code")
